+++ tools.go
@@ -1,8 +1,0 @@
-// +build tools
-// This file only exists so that tools we need for go generate can be versioned in our go.mod
-
-package linstor_csi
-
-import (
-	_ "github.com/vektra/mockery/v2"
-)